edited build script and removed unnecessary files

+++ 7.2/stretch/apache/scripts/build-scrambled.sh
@@ -1,13 +1,17 @@
 #!/bin/bash
 
+cd $POLYSCRIPT_PATH
+
 ./php-scrambler
-
-s_php tok-php-transformer.php -p "$PHP_SRC_PATH"ext/phar --replace --inc --phar
-
-cd $PHP_SRC_PATH
-make install -k
+if [ -f ./scrambled.json ]; then 
+        s_php tok-php-transformer.php -p "$PHP_SRC_PATH"ext/phar --replace --inc --phar
+        cd $PHP_SRC_PATH
+        make install -k
+else 
+        exit 1
+fi
 
 cd $POLYSCRIPT_PATH
 s_php tok-php-transformer.php -p "$PHP_SRC_PATH"ext/phar/phar.php --replace
 cd $PHP_SRC_PATH
-make install -k
+make pharcmd -k
